+++ src/app/components/alerts/alerts.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { HotToastService } from '@ngneat/hot-toast';
+import { AdminService } from 'src/app/services/admin.service';
 
 @Component({
   selector: 'app-alerts',
@@ -7,63 +8,108 @@ import { HotToastService } from '@ngneat/hot-toast';
   styleUrls: ['./alerts.component.scss']
 })
 export class AlertsComponent implements OnInit {
-
+  fetchedActiveAlerts: any = [];
   constructor(
-    private toastService: HotToastService
+    private toastService: HotToastService,
+    private adminApi: AdminService
   ) { }
 
   ngOnInit(): void {
-    this.apiError(false);
-    this.betaTesting(false);
-    this.registrationOpen(false);
+    this.adminApi.adminGetActiveAlerts().subscribe(res => {
+      this.fetchedActiveAlerts = res;
+      this.getActiveAlerts(this.fetchedActiveAlerts);
+    });
  }
 
-  apiError(bool: boolean){
-    if(bool == true){
-       this.toastService.error('There is currently a problem with the authorization API. <br/><b>AWE may not work as expected.<b>', {
-        style: {
-          width: '100%',
-          textAlign: 'center',
-        },
-        theme: 'snackbar',
-        position: 'top-center',
-        autoClose: false,
-        id: 'auth-api-error-8-13-2021',
-        persist: { enabled: false, count: 1 },
-      });
-    }
+ getActiveAlerts(alerts: any){
+   var i = 1;
+   if(alerts){
+   for(let alert of alerts){
+     alert.persist = Boolean(JSON.parse(alert.persist));
+     alert.autoclose = Boolean(JSON.parse(alert.autoclose));
+     if(alert.persist_id == '' || !alert.persist_id){
+       alert.persist_id = 'no-id-' + i;
+     }
+       if(alert.type == 'show'){
+         this.toastService.show(alert.message, {
+          style: {
+            width: '100%',
+            textAlign: 'center',
+          },
+          theme: alert.theme,
+          position: 'top-center',
+          autoClose: alert.autoclose,
+          id: alert.persist_id,
+          persist: { enabled: alert.persist, count: alert.persist_count },
+          });
+        }
+        else if(alert.type == 'info'){
+          this.toastService.info(alert.message, {
+           style: {
+             width: '100%',
+             textAlign: 'center',
+           },
+           theme: alert.theme,
+           position: 'top-center',
+           autoClose: alert.autoclose,
+           id: alert.persist_id,
+           persist: { enabled: alert.persist, count: alert.persist_count },
+         });
+       }
+       else if(alert.type == 'success'){
+         this.toastService.success(alert.message, {
+          style: {
+            width: '100%',
+            textAlign: 'center',
+          },
+          theme: alert.theme,
+          position: 'top-center',
+          autoClose: alert.autoclose,
+          id: alert.persist_id,
+          persist: { enabled: alert.persist, count: alert.persist_count },
+          });
+        }
+        else if(alert.type == 'warning'){
+          this.toastService.warning(alert.message, {
+           style: {
+             width: '100%',
+             textAlign: 'center',
+           },
+           theme: alert.theme,
+           position: 'top-center',
+           autoClose: alert.autoclose,
+           id: alert.persist_id,
+           persist: { enabled: alert.persist, count: alert.persist_count },
+         });
+       }
+       else if(alert.type == 'error'){
+         this.toastService.error(alert.message, {
+          style: {
+            width: '100%',
+            textAlign: 'center',
+          },
+          theme: alert.theme,
+          position: 'top-center',
+          autoClose: alert.autoclose,
+          id: alert.persist_id,
+          persist: { enabled: alert.persist, count: alert.persist_count },
+        });
+      }
+      else{
+        this.toastService.show(alert.message, {
+         style: {
+           width: '100%',
+           textAlign: 'center',
+         },
+         theme: alert.theme,
+         position: 'top-center',
+         autoClose: alert.autoclose,
+         id: alert.persist_id,
+         persist: { enabled: alert.persist, count: alert.persist_count },
+       });
+     }
+     i = i + 1;
+   }
   }
-
-  betaTesting(bool: boolean){
-    if(bool == true){
-       this.toastService.info('<b>BETA TESTING UPDATE</b><br/>We are currently in the proccess of prepping AWE for beta testing.<br/><i>Stay Tuned!</i>', {
-        style: {
-          width: '100%',
-          textAlign: 'center',
-        },
-        theme: 'snackbar',
-        position: 'top-center',
-        autoClose: false,
-        id: 'beta-testing-8-11-2021',
-        persist: { enabled: false, count: 1 },
-      });
-    }
-  }
-
-  registrationOpen(bool: boolean){
-    if(bool == true){
-       this.toastService.success('<b>REGISTRATION IS OPEN</b><br/>We have enabled registration!<br/>Feel free to create an account and have a look around AWE v2.<br/>We hope you enjoy it!', {
-        style: {
-          width: '100%',
-          textAlign: 'center',
-        },
-        theme: 'snackbar',
-        position: 'top-center',
-        autoClose: false,
-        id: 'beta-testing-8-11-2021',
-        persist: { enabled: false, count: 1 },
-      });
-    }
-  }
-
+ }
 }